+++ tests/JoliTypo/Tests/JoliTypoTest.php
@@ -135,11 +135,21 @@ class JoliTypoTest extends \PHPUnit_Framework_TestCase
 
         $this->assertEquals("Mentions L&eacute;gales", $fixer->fix(utf8_encode(utf8_decode("Mentions Légales"))));
 
-        // JoliTypo can't handle double encoded UTF-8 strings, nor ISO strings
+        // JoliTypo can handle double encoded UTF-8 strings, or ISO strings, but that's not a feature.
         $isoString = mb_convert_encoding("Mentions Légales", "ISO-8859-1", "UTF-8");
         $this->assertEquals("Mentions L&eacute;gales", $fixer->fix(utf8_encode($isoString)));
-        $this->assertNotEquals("Mentions L&eacute;gales", $fixer->fix($isoString));
+        $this->assertEquals("Mentions L&eacute;gales", $fixer->fix($isoString));
         $this->assertEquals("Mentions L&Atilde;&copy;gales", $fixer->fix(utf8_encode(utf8_encode($isoString))));
+    }
+
+    public function testEmptyContent()
+    {
+        $fixer = new Fixer(array('Trademark'));
+        $this->assertInstanceOf('JoliTypo\Fixer', $fixer);
+
+        $this->assertEquals("", $fixer->fix(""));
+        $this->assertEquals("\n ", $fixer->fix("\n "));
+        $this->assertEquals("some content", $fixer->fix("\n some content"));
     }
 }